完善技能战斗ing 完善奖励发放ing

+++ src/main/java/com/linlazy/mmorpg/module/common/reward/Reward.java
@@ -12,7 +12,7 @@ public class Reward {
      * 奖励ID
      */
 
-    private long rewardId;
+    private int rewardId;
     /**
      * 奖励数量
      */
@@ -26,7 +26,7 @@ public class Reward {
     public Reward() {
     }
 
-    public Reward(long rewardId, int count, int rewardDBType) {
+    public Reward(int rewardId, int count, int rewardDBType) {
         this.rewardId = rewardId;
         this.count = count;
         this.rewardDBType = rewardDBType;
